// sourcecode/BTrak.Packages/snova-project-management/PM-module-components/pm-components/src/lib/snovasys-projects/models/userStory.ts
export class UserStory {
  [x: string]: any;
  id: string;
  userStory: string;
  userStoryId: string;
  parentUserStoryId: string;
  parentUserStoryGoalId: string;
  userStoryName: string;
  projectId: string;
  goalId: string;
  goalIdForBug: string;

  estimatedTime: any;
// ...
}